# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TourFlow AI
  description: 'TourFlow AI API contract — generated from the FastAPI implementation
    (source of truth: backend/app). Regenerate: python -c "from app.main import app;
    ..."'
  version: 1.0.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - users
      summary: Me
      operationId: me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileOut'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - users
      summary: Update Me
      operationId: update_me_api_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/users:
    get:
      tags:
      - users
      summary: List Users
      operationId: list_users_api_v1_users_get
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Role'
          - type: 'null'
          title: Role
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileOut'
                title: Response List Users Api V1 Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id}/role:
    patch:
      tags:
      - users
      summary: Set Role
      operationId: set_role_api_v1_users__user_id__role_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications:
    get:
      tags:
      - users
      summary: My Notifications
      operationId: my_notifications_api_v1_notifications_get
      security:
      - HTTPBearer: []
      parameters:
      - name: unread_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Unread Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationOut'
                title: Response My Notifications Api V1 Notifications Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/{nid}/read:
    post:
      tags:
      - users
      summary: Mark Read
      operationId: mark_read_api_v1_notifications__nid__read_post
      security:
      - HTTPBearer: []
      parameters:
      - name: nid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Nid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/audit-logs:
    get:
      tags:
      - users
      summary: Audit Logs
      operationId: audit_logs_api_v1_audit_logs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditOut'
                title: Response Audit Logs Api V1 Audit Logs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/reviews:
    post:
      tags:
      - users
      summary: Create Review
      operationId: create_review_api_v1_trips__trip_id__reviews_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - users
      summary: List Reviews
      operationId: list_reviews_api_v1_trips__trip_id__reviews_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewOut'
                title: Response List Reviews Api V1 Trips  Trip Id  Reviews Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operator/summary:
    get:
      tags:
      - users
      summary: Operator Summary
      operationId: operator_summary_api_v1_operator_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
components:
  schemas:
...
    AuditOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        actor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Actor Id
        action:
          type: string
          title: Action
        resource_type:
          type: string
          title: Resource Type
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        old_state:
          type: object
          title: Old State
        new_state:
          type: object
          title: New State
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - action
      - resource_type
      - old_state
      - new_state
      - created_at
      title: AuditOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
        kind:
          type: string
          title: Kind
        read:
          type: boolean
          title: Read
        trip_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Trip Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - title
      - kind
      - read
      - created_at
      title: NotificationOut
...
    ProfileOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        role:
          $ref: '#/components/schemas/Role'
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
      type: object
      required:
      - id
      - email
      - role
      title: ProfileOut
    ProfileUpdate:
      properties:
        full_name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Full Name
        phone:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Phone
      type: object
      title: ProfileUpdate
...
    ReviewIn:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
      type: object
      required:
      - rating
      title: ReviewIn
    ReviewOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        traveler_id:
          type: string
          format: uuid
          title: Traveler Id
        rating:
          type: integer
          title: Rating
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - trip_id
      - traveler_id
      - rating
      - created_at
      title: ReviewOut
    Role:
      type: string
      enum:
      - TRAVELER
      - OPERATOR
      - COORDINATOR
      - VENDOR
      - ADMIN
      title: Role
    RoleUpdate:
      properties:
        role:
          $ref: '#/components/schemas/Role'
      type: object
      required:
      - role
      title: RoleUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer